Accounts Receivable

Audit sample obtained from a population of N = 87 accounts receivable, totaling $612,824 in book value (Higgins and Nandram, 2009; Lohr, 2021).

data(accounts)

Format

A data frame with 20 rows and 3 variables.

account

account number (between 1 - 87)

bookValue

booked value of the account

auditValue

audited (i.e., true) value of the account

References

Higgins, H. N., & Nandram, B. (2009). Monetary unit sampling: Improving estimation of the total audit error Advances in Accounting, 25(2), 174-182. doi:10.1016/j.adiac.2009.06.001

Lohr, S. L. (2021). Sampling: Design and Analysis. CRC press.

Examples

data(accounts)