Legitimacy Audit

Fictional population from a legitimacy audit, containing 4189 records with identification numbers, stratum identifiers, book values, and audit values. The audit values are added for illustrative purposes, as these would need to be assessed by the auditor in the execution stage of the audit.

Usage

data(allowances)

Format

A data frame with 4189 rows and 5 variables.

item

a unique record identification number.

branch

the stratum identifyer / branch number.

bookValue

the item book value in US dollars.

auditValue

the item audit (true) value in US dollars.

times

a sample selection indicator (0 = not in sample).

Examples

data(allowances)