Auditing Data: Digit Distribution Test

This function extracts and performs a test of the distribution of (leading) digits in a vector against a reference distribution. By default, the distribution of leading digits is checked against Benford's law.

Usage

digit_test(x, check = c("first", "last", "firsttwo"),
           reference = 'benford', prior = FALSE)

Arguments

x

a numeric vector.

check

location of the digits to analyze. Can be first, firsttwo, or last.

reference

which character string given the reference distribution for the digits, or a vector of probabilities for each digit. Can be benford for Benford's law, uniform for the uniform distribution. An error is given if any entry of reference is negative. Probabilities that do not sum to one are normalized.

prior

a logical specifying whether to use a prior distribution, or a numeric vector containing the prior parameters for the Dirichlet distribution on the digit categories.

Value

An object of class jfaDistr containing:

observed

the observed counts.

expected

the expected counts under the null hypothesis.

n

the number of observations in x.

statistic

the value the chi-squared test statistic.

parameter

the degrees of freedom of the approximate chi-squared distribution of the test statistic.

p.value

the p-value for the test.

check

checked digits.

digits

vector of digits.

reference

reference distribution

data.name

a character string giving the name(s) of the data.

Details

Benford's law is defined as \(p(d) = log10(1/d)\). The uniform distribution is defined as \(p(d) = 1/d\).

References

Benford, F. (1938). The law of anomalous numbers. In Proceedings of the American Philosophical Society, 551-572.

See also

Author

Koen Derks, k.derks@nyenrode.nl

Examples

set.seed(1)
x <- rnorm(100)

# First digit analysis against Benford's law
digit_test(x, check = "first", reference = "benford")
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, X-squared = 14.557, df = 8, p-value = 0.06836
#> alternative hypothesis: leading digit(s) are not distributed according to the benford distribution.

# Bayesian first digit analysis against Benford's law
digit_test(x, check = "first", reference = "benford", prior = TRUE)
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, BF10 = 0.019696
#> alternative hypothesis: leading digit(s) are not distributed according to the benford distribution.

# Last digit analysis against the uniform distribution
digit_test(x, check = "last", reference = "uniform")
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, X-squared = 4.22, df = 8, p-value = 0.8367
#> alternative hypothesis: last digit(s) are not distributed according to the uniform distribution.

# Bayesian last digit analysis against the uniform distribution
digit_test(x, check = "last", reference = "uniform", prior = TRUE)
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, BF10 = 0.00018458
#> alternative hypothesis: last digit(s) are not distributed according to the uniform distribution.

# First digit analysis against a custom distribution
digit_test(x, check = "last", reference = 1:9)
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, X-squared = 86.689, df = 8, p-value = 2.182e-15
#> alternative hypothesis: last digit(s) are not distributed according to the reference distribution.

# Bayesian first digit analysis against a custom distribution
digit_test(x, check = "last", reference = 1:9, prior = TRUE)
#> 
#> 	Digit distribution test
#> 
#> data:  x
#> n = 100, BF10 = 1970104
#> alternative hypothesis: last digit(s) are not distributed according to the reference distribution.