Methods for jfa objects

Methods defined for objects returned from the auditPrior, planning, selection, and evaluation functions.

# S3 method for jfaPrior
print(x, ...)

# S3 method for summary.jfaPrior
print(x, digits = getOption("digits"), ...)

# S3 method for jfaPrior
summary(object, digits = getOption("digits"), ...)

# S3 method for jfaPrior
predict(object, n, lim = NULL, cumulative = FALSE, ...)

# S3 method for jfaPrior
plot(x, xlim = c(0, 1), ...)

# S3 method for jfaPosterior
print(x, ...)

# S3 method for summary.jfaPosterior
print(x, digits = getOption("digits"), ...)

# S3 method for jfaPosterior
summary(object, digits = getOption("digits"), ...)

# S3 method for jfaPosterior
predict(object, n, lim = NULL, cumulative = FALSE, ...)

# S3 method for jfaPosterior
plot(x, xlim = c(0, 1), ...)

# S3 method for jfaPlanning
print(x, ...)

# S3 method for summary.jfaPlanning
print(x, digits = getOption("digits"), ...)

# S3 method for jfaPlanning
summary(object, digits = getOption("digits"), ...)

# S3 method for jfaPlanning
plot(x, xlim = c(0, 1), ...)

# S3 method for jfaSelection
print(x, ...)

# S3 method for summary.jfaSelection
print(x, digits = getOption("digits"), ...)

# S3 method for jfaSelection
summary(object, digits = getOption("digits"), ...)

# S3 method for jfaSelection
plot(x, ...)

# S3 method for jfaEvaluation
print(x, digits = getOption("digits"), ...)

# S3 method for summary.jfaEvaluation
print(x, digits = getOption("digits"), ...)

# S3 method for jfaEvaluation
summary(object, digits = getOption("digits"), ...)

# S3 method for jfaEvaluation
plot(x, xlim = c(0, 1), ...)

# S3 method for jfaDistr
print(x, digits = getOption("digits"), ...)

# S3 method for jfaDistr
plot(x, ...)

# S3 method for jfaRv
print(x, digits = getOption("digits"), ...)

# S3 method for jfaRv
plot(x, ...)

Arguments

...

further arguments, currently ignored.

digits

an integer specifying the number of digits to which output should be rounded. Used in summary.

object, x

an object of class jfaPrior, jfaPosterior, jfaPlanning, jfaSelection, jfaEvaluation, jfaDistr, or jfaRv.

n

used in predict. Specifies the sample size for which predictions should be made.

lim

used in predict. Limits the number of errors for which predictions should be made.

cumulative

used in predict. Specifies whether cumulative probabilities should be shown.

xlim

used in plot. Specifies the x limits (x1, x2) of the plot.

Value

The summary methods return a data.frame which contains the input and output.

The print methods simply print and return nothing.