Algorithm Auditing: Fairness Metrics and Bias Detection

This function detects bias in algorithmic decision-making systems by computing various fairness metrics. It provides an assessment of fairness across different groups based on the observed and predicted outcomes. The function allows for evaluating fairness using metrics such as demographic parity, proportional parity, predictive rate parity, accuracy parity, false negative rate parity, false positive rate parity, true positive rate parity, negative predicted value parity, and specificity parity and decide whether groups are fair to a certain degree and within a certain materiality threshold. Currently, it only supports binary classification.

model_fairness(
  data,
  sensitive,
  target,
  predictions,
  reference = NULL,
  positive = NULL,
  materiality = 0.2,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)

Arguments

data

The input data.

sensitive

The column name indicating the sensitive variable.

target

The column name indicating the target class labels.

predictions

The column name indicating the predictions class labels.

reference

The reference class for computing the fairness metrics. If NULL (the default), the first level of the sensitive column is used.

positive

The positive class for computing the fairness metrics. If NULL (the default), the first level of the target column is used.

materiality

The materiality value for determining fairness.

alternative

The type of confidence interval to produce. Possible options are two.sided (the default), greater and less.

conf.level

a numeric value between 0 and 1 specifying the confidence level (i.e., 1 - audit risk / detection risk).

Value

An object of class jfaModelBias containing:

reference

The reference group for computing the fairness metrics.

positive

The positive class used in computing the fairness metrics.

alternative

The type of confidence interval.

confusion.matrix

A list of confusion matrices for each group.

performance

A data frame containing performance metrics for each group, including accuracy, precision, recall, and F1 score.

fairness

A data frame containing fairness metrics for each group, including demographic parity, proportional parity, predictive rate parity, accuracy parity, false negative rate parity, false positive rate parity, true positive rate parity, negative predicted value parity, and statistical parity.

ratio

A data frame containing fairness ratios for each metric, comparing each group to the reference group.

materiality

The materiality value used to determine the out of bounds metrics.

data.name

The name of the input data object.

Details

The following fairness metrics are computed on the basis of the true positives (TP), false positives (FP), true negative (TN) and false negatives (FN) in the confusion matrix for each group.

Demographic parity: measures whether the observed variable is distributed equally across different groups, calculated as TP + FP.

Proportional parity: measures whether the positive rate is distributed equally across different groups, calculated as (TP + FP) / (TP + FP + TN + FN).

Predictive rate parity: measures whether the positive prediction rate is the same across different groups, calculated as TP / (TP + FP).

Accuracy parity: measures whether the overall accuracy is the same across different groups, calculated as (TP + TN) / (TP + FP + TN + FN).

False negative rate parity: measures whether the false negative rate is the same across different groups, calculated as FN / (FP + FN).

False positive rate parity: measures whether the false positive rate is the same across different groups, calculated as FP / (TN + FP).

True positive rate parity: measures whether the true positive rate is the same across different groups, calculated as TP / (TP + FN).

Negative predicted value parity: measures whether the negative predicted value is the same across different groups, calculated as TN / (TN + FN).

Specificity parity: measures whether the true positive rate is the same across different groups, calculated as TN / (TN + FP).

References

Pessach, D. & Shmueli, E. (2022). A review on fairness in machine learning. ACM Computing Surveys, 55(3), 1-44. doi:10.1145/3494672

Author

Koen Derks, k.derks@nyenrode.nl

Examples

model_fairness(compas, "Ethnicity", "TwoYrRecidivism", "Predicted",
  reference = "Caucasian", positive = "yes"
)
#> 
#> 	Fairness Metrics and Bias Detection
#> 
#> data:  compas
#> reference group:  Caucasian, positive class:  yes
#> 
#> Sensitive groups (5):
#>  African_American: 6 measures outside tolerance region
#>  Asian: 5 measures outside tolerance region
#>  Hispanic: 1 measures outside tolerance region
#>  Native_American: 5 measures outside tolerance region
#>  Other: 3 measures outside tolerance region