report {jfa}	R Documentation

Generate an Audit Report

Description

This function takes an object of class jfaEvaluation, creates a report containing the results, and saves the report to a file in your working directory.

For more details on how to use this function see the package vignette: vignette("jfa", package = "jfa")

Usage


report(object = NULL, file = NULL, format = "html_document")

Arguments

object
an object of class 'jfaEvaluation' as returned by the evaluation() function.

file
a string that gives the desired name of the file (e.g. "report.html"). The report is created in your current working directory.

format
can be either one of "html_document" or "pdf_document" (required MikTex).

Value

A html or pdf report containing the results of the evaluation.

Author(s)

Koen Derks, k.derks@nyenrode.nl

See Also

evaluation

Examples


library(jfa)
set.seed(1)

# Generate some audit data (N = 1000):
data <- data.frame(ID = sample(1000:100000, size = 1000, replace = FALSE), 
                   bookValue = runif(n = 1000, min = 700, max = 1000))

# Using monetary unit sampling, draw a random sample from the population.
s1 <- selection(population = data, sampleSize = 100, units = "mus", 
                 bookValues = "bookValue", algorithm = "random")
s1_sample <- s1$sample
s1_sample$trueValue <- s1_sample$bookValue
s1_sample$trueValue[2] <- s1_sample$trueValue[2] - 500 # One overstatement is found

e2 <- evaluation(sample = s1_sample, bookValues = "bookValue", auditValues = "trueValue", 
                 method = "stringer", materiality = 0.05, counts = s1_sample$counts)

# Generate report
# report(e2, file = "myFile.html")


[Package jfa version 0.5.0 Index]