auditPrior {jfa}	R Documentation

Create a Prior Distribution with Audit Information

Description

This function creates a prior distribution for Bayesian audit sampling according to several methods discussed in Derks et al. (2020). The returned object is of class jfaPrior and can be used with associated print() and plot() methods. jfaPrior objects can be used as input for the prior argument in other functions.

Usage


auditPrior(materiality = NULL, confidence = 0.95, method = "arm", ir = 1, 
           cr = 1, expectedError = 0, likelihood = "binomial", N = NULL, 
           pHmin = NULL, pHplus = NULL, factor = 1, sampleN = 0, sampleK = 0)

Arguments

materiality
a value between 0 and 1 representing the materiality of the audit as a fraction of the total size or value. Can be NULL for some methods.

confidence
the confidence level desired from the confidence bound (on a scale from 0 to 1). Defaults to 0.95, or 95% confidence.

method
the method by which the prior distribution is constructed. Defaults to the arm method, which uses the audit risk model (Derks et al., 2019). Can be one of none, median, hypotheses, arm, sample or factor. See the Details section for more information.

ir
the inherent risk probability from the audit risk model. Defaults to 1 for 100% risk.

cr
the inherent risk probability from the audit risk model. Defaults to 1 for 100% risk.

expectedError
a fraction representing the percentage of expected mistakes in the sample relative to the total size, or a number (>= 1) that represents the number of expected mistakes.

likelihood
can be one of binomial, poisson, or hypergeometric. See the Details section for more information.

N
the population size (only required when likelihood = 'hypergeometric').

pHmin
When using method = 'hypotheses', the prior probability of the hypothesis θ < materiality.

pHplus
When using method = 'hypotheses', the prior probability of the hypothesis θ > materiality.

factor
When using method = 'factor', the value of the weighting factor for the results of the previous sample.

sampleN
When using method sample or factor, the number of transactions that were inspected in the previous sample.

sampleK
When using method sample or factor, the total taint in the previous sample.

Details

This section elaborates on the available methods for constructing a prior distribution.

none: This method constructs a prior distribution according to the principle of minimum information.

median: This method constructs a prior distribution so that the prior probabilities of tolerable and intolerable misstatement are equal.

hypotheses: This method constructs a prior distribution with specified prior probabilities for the hypotheses of tolerable and intolerable misstatement. Requires specification of the pHmin and pHplus arguments.

arm: This method constructs a prior distribution according to the assessed risks in the audit risk model. Requires specification of the ir and cr arguments.

sample: This method constructs a prior distribution on the basis of an earlier sample. Requires specification of the sampleN and sampleK arguments.

factor: This method constructs a prior distribution on the basis of last year's results and a weighting factor. Requires specification of the factor, sampleN and sampleK arguments.

This section elaborates on the available likelihoods and corresponding prior distributions for the likelihood argument.

poisson: The Poisson likelihood is used as a likelihood for monetary unit sampling (MUS). Its likelihood function is defined as:

p(x) = \frac{λ^x e^{-λ}}{x!}

The conjugate gamma(α, β) prior has probability density function:

f(x; α, β) = \frac{β^α x^{α - 1} e^{-β x}}{Γ(α)}

binomial: The binomial likelihood is used as a likelihood for record sampling with replacement. Its likelihood function is defined as:

p(x) = {n \choose k} p^k (1 - p)^{n - k}

The conjugate beta(α, β) prior has probability density function:

f(x; α, β) = \frac{1}{B(α, β)} x^{α - 1} (1 - x)^{β - 1}

hypergeometric: The hypergeometric likelihood is used as a likelihood for record sampling without replacement. Its likelihood function is defined as:

p(x = k) = \frac{{K \choose k} {N - K \choose n - k}}{{N \choose n}}

The conjugate beta-binomial(α, β) prior (Dyer and Pierce, 1993) has probability density function:

f(k | n, α, β) = {n \choose k} \frac{B(k + α, n - k + β)}{B(α, β)}

Value

An object of class jfaPrior containing:

method
the method by which the prior distribution is constructed.

likelihood
the likelihood by which the prior distribution is updated.

priorD
the name of the probability density function of the prior distribution.

nPrior
the prior assumed sample size.

kPrior
the prior assumed sample errors

aPrior
the prior parameter alpha.

bPrior
the prior parameter beta.

materiality
the materiality that was used to construct the prior distribution.

N
if specified as input, the population size.

pHmin
For methods median and hypotheses, the prior probability of the hypothesis θ < materiality.

pHmin
For methods median and hypotheses, the prior probability of the hypothesis θ > materiality.

sampleN
For methods sample and factor, the total number of transactions in the earlier sample.

sampleK
For methods sample and factor, the number of transactions that were misstated in the earlier sample.

Author(s)

Koen Derks, k.derks@nyenrode.nl

References

Derks, K., de Swart, J., Wagenmakers, E.-J., Wille, J., & Wetzels, R. (2019). JASP for audit: Bayesian tools for the auditing practice.

Derks, K., de Swart, J., van Batenburg, P. Wagenmakers, E.-J., & Wetzels, R. (2020). Priors in a Bayesian Audit: How Integrating Information into the Prior Distribution can Improve Audit Transparency and Efficiency.

See Also

planning sampling evaluation

Examples


library(jfa)

# Specify the materiality, confidence, and expected errors:
materiality   <- 0.05   # 5%
confidence    <- 0.95   # 95%
expectedError <- 0.025  # 2.5%

# Specify the inherent risk (ir) and control risk (cr):
ir <- 1     # 100%
cr <- 0.6   # 60%

# Create a beta prior distribution according to the Audit Risk Model (arm) 
# and a binomial likelihood:
prior <- auditPrior(materiality = materiality, confidence = confidence, 
                    method = "arm", ir = ir, cr = cr, 
                    expectedError = expectedError, likelihood = "binomial")
print(prior)

# ------------------------------------------------------------
#         jfa Prior Distribution Summary (Bayesian)
# ------------------------------------------------------------
# Input:
# 
# Confidence:              0.95    
# Expected sample errors:  0.025       
# Likelihood:              binomial 
# Specifics:               Inherent risk = 1; Internal control risk = 0.6; Detection risk = 0.083 
# ------------------------------------------------------------
# Output: 
# 
# Prior distribution:      beta(2.275, 50.725) 
# Implicit sample size:    51 
# Implicit errors:         1.275 
# ------------------------------------------------------------

[Package jfa version 0.3.1 Index]