Create a prior distribution for audit sampling

This function creates a prior distribution with audit information to be used in the planning() and evaluation() functions via their prior argument. The function returns an object of class jfaPrior which can be used with associated print() and plot() methods.

For more details on how to use this function, see the package vignette: vignette('jfa', package = 'jfa')

auditPrior(confidence, materiality = NULL, expectedError = 0, 
           method = 'none', likelihood = 'binomial', N = NULL, 
           ir = 1, cr = 1, ub = NULL, pHmin = NULL, pHplus = NULL, 
           sampleN = 0, sampleK = 0, factor = 1)

Arguments

confidence
a numeric value between 0 and 1 specifying the confidence level to be used in the planning.

materiality
a numeric value between 0 and 1 specifying the performance materiality (i.e., the maximum upper limit) as a fraction of the total population size. Can be NULL for some methods.

expectedError
a numeric value between 0 and 1 specifying the expected errors in the sample relative to the total sample size, or a numeric value (>= 1) that represents the sum of expected errors in the sample. It is advised to set this value conservatively to minimize the probability of the observed errors exceeding the expected errors, which would imply that insufficient work has been done in the end.

method
a character specifying the method by which the prior distribution is constructed. Defaults to none which incorporates no existing information. Other options are arm, bram, median, hypotheses, sample, and factor. See the details section for more information about the available methods.

likelihood
a character specifying the likelihood assumed when updating the prior distribution. This can be either binomial for the binomial likelihood and beta prior distribution, poisson for the Poisson likelihood and gamma prior distribution, or hypergeometric for the hypergeometric likelihood and beta-binomial prior distribution. See the details section for more information about the available likelihoods.

N
an numeric value larger than 0 specifying the total population size. Optional unless likelihood = 'hypergeometric'.

ir
if method = 'arm', a numeric value between 0 and 1 specifying the inherent risk in the audit risk model. Defaults to 1 for 100% risk.

cr
if method = 'arm', a numeric value between 0 and 1 specifying the internal control risk in the audit risk model. Defaults to 1 for 100% risk.

ub
if method = 'bram', a numeric value between 0 and 1 specifying the upper bound for the prior distribution as a fraction of the population size.

pHmin
if method = 'hypotheses', a numeric value between 0 and 1 specifying the prior probability of the hypothesis \(\theta <\) materiality.

pHplus
if method = 'hypotheses', a numeric value between 0 and 1 specifying the prior probability of the hypothesis \(\theta >\) materiality.

sampleN
if method = 'sample' or method = 'factor', an integer larger than, or equal to, 0 specifying the sample size of the sample equivalent to the prior information.

sampleK
if sample or factor, a numeric value larger than, or equal to, 0 specifying the sum of errors in the sample equivalent to the prior information.

factor
if method = 'factor', a numeric value between 0 and 1 specifying the weighting factor for the results of the sample equivalent to the prior information.

Value

An object of class jfaPrior containing:

confidence

a numeric value between 0 and 1 indicating the confidence level used.

materiality

if materiality is specified, a numeric value between 0 and 1 indicating the materiality used to construct the prior distribution.

expectedError

a numeric value larger than, or equal to, 0 indicating the input for the number of expected errors.

method

a character indicating the method by which the prior distribution is constructed.

likelihood

a character indicating the assumed likelihood.

N

if N is specified, an integer larger than 0 indicating the population size.

description

a list containing a description of the prior distribution, including the parameters of the prior distribution and the implicit sample on which the prior distribution is based.

statistics

a list containing statistics of the prior distribution, including the mean, mode, median, and upper bound of the prior distribution.

specifics

a list containing specifics of the prior distribution that vary depending on the method.

hypotheses

if materiality is specified, a list containing information about the hypotheses, including prior probabilities and odds for the hypothesis of tolerable misstatement (H-) and the hypothesis of intolerable misstatement (H+).

Details

This section elaborates on the available options for the method argument.

none: This method constructs a prior distribution that incorporates negligible information about the possible values of the misstatement.

arm: This method constructs a prior distribution by translating the risks of material misstatement (inherent risk and internal control risk) from the audit risk model to an implicit sample. The method requires specification of the ir (inherent risk) and cr (internal control risk) arguments.

bram: This method constructs a prior distribution using the Bayesian audit risk assessment model (BRAM) in which the expected most likely error and expected upper bound of the misstatement must be specified. The method requires specification of the ub argument.

median: This method constructs a prior distribution so that the prior probability of tolerable misstatement (\(\theta <\) materiality) is equal to the prior probability of intolerable misstatement (\(\theta >\) materiality).

hypotheses: This method constructs a prior distribution with custom prior probabilities for the hypotheses of tolerable misstatement (\(\theta <\) materiality) and intolerable misstatement (\(\theta >\) materiality). This method requires specification of the pHmin and pHplus arguments.

sample: This method constructs a prior distribution on the basis of an earlier sample. This method requires specification of the sampleN and sampleK arguments.

factor: This method constructs a prior distribution on the basis of an earlier sample in combination with a weighting factor. This method requires specification of the sampleN, sampleK, and factor arguments.

This section elaborates on the available likelihoods and corresponding prior distributions for the likelihood argument.

binomial: The binomial likelihood is often used as a likelihood for attributes sampling with replacement. The likelihood function is defined as: $$p(x) = {n \choose k} p^k (1 - p)^{n - k}$$ The conjugate beta(\(\alpha, \beta\)) prior has probability density function: $$f(x; \alpha, \beta) = \frac{1}{B(\alpha, \beta)} x^{\alpha - 1} (1 - x)^{\beta - 1}$$

poisson: The Poisson likelihood is often used as a likelihood for monetary unit sampling (MUS). The likelihood function is defined as: $$p(x) = \frac{\lambda^x e^{-\lambda}}{x!}$$ The conjugate gamma(\(\alpha, \beta\)) prior has probability density function: $$f(x; \alpha, \beta) = \frac{\beta^\alpha x^{\alpha - 1} e^{-\beta x}}{\Gamma(\alpha)}$$

hypergeometric: The hypergeometric likelihood is used as a likelihood for sampling without replacement. The likelihood function is defined as: $$p(x = k) = \frac{{K \choose k} {N - K \choose n - k}}{{N \choose n}}$$ The conjugate beta-binomial(\(\alpha, \beta\)) prior (Dyer and Pierce, 1993) has probability density function: $$f(k | n, \alpha, \beta) = {n \choose k} \frac{B(k + \alpha, n - k + \beta)}{B(\alpha, \beta)}$$

References

Derks, K., de Swart, J., Wagenmakers, E.-J., Wille, J., & Wetzels, R. (2019). JASP for audit: Bayesian tools for the auditing practice.

Derks, K., de Swart, J., van Batenburg, P. Wagenmakers, E.-J., & Wetzels, R. (2021). Priors in a Bayesian audit: How integrating information into the prior distribution can improve audit transparency and efficiency. In Press.

See also

Author

Koen Derks, k.derks@nyenrode.nl

Examples

 
# Translate inherent risk (ir) and control risk (cr) to a prior distribution
auditPrior(confidence = 0.95, materiality = 0.05, expectedError = 0.025,
           method = 'arm', likelihood = 'binomial', ir = 1, cr = 0.6)

#> # ------------------------------------------------------------
#> #         jfa Prior Distribution Summary (Bayesian)
#> # ------------------------------------------------------------
#> # Input:
#> #
#> # Confidence:                    95% 
#> # Expected sample errors:        2.5% 
#> # Likelihood:                    binomial 
#> # Specifics:                     Inherent risk = 1; Internal control risk = 0.6; Detection risk = 0.083 
#> # ------------------------------------------------------------
#> # Output: 
#> #
#> # Prior distribution:            beta(α = 2.275, β = 50.725) 
#> # Implicit sample size:          51 
#> # Implicit errors:               1.275 
#> # ------------------------------------------------------------
#> # Statistics: 
#> #
#> # Upper bound:                   0.096 
#> # Precision:                     0.071 
#> # Mode:                          0.025 
#> # Mean:                          0.043 
#> # Median:                        0.037 
#> # ------------------------------------------------------------