jfa --- Bayesian and Classical Audit Sampling

jfa is an R package for statistical audit sampling. The package provides functions for planning, performing, evaluating, and reporting an audit sample. Specifically, these functions implement standard audit sampling techniques for calculating sample sizes, selecting items from a population, and evaluating the misstatement from a data sample or from summary statistics. Additionally, the jfa package allows the user to create a prior probability distribution to perform Bayesian audit sampling using these functions.

The package and its intended workflow are also implemented with a graphical user interface in the Audit module of JASP, a free and open-source statistical software program.

For documentation on jfa itself, including the manual and user guide for the package, worked examples, and other tutorial information visit the package website.

Reference tables

Below you can find several links to reference tables that contain statistical sample sizes, upper limits, and Bayes factors. These tables are created using the planning() and evaluation() functions provided in the package. See the corresponding help files for more information about these functions and how to replicate this output.

Sample sizes

Sample sizes based on the Poisson distribution

Sample sizes based on the binomial distribution

Sample sizes based on the hypergeometric distribution

Upper limits

Upper limits based on the Poisson distribution

Upper limits based on the binomial distribution

Upper limits based on the hypergeometric distribution

One-sided p values

One sided p values based on the Poisson distribution

One sided p values based on the binomial distribution

One sided p values based on the hypergeometric distribution

Bayes factors

Impartial Bayes factors based on the gamma distribution

Impartial Bayes factors based on the beta distribution

Impartial Bayes factors based on the beta-binomial distribution

Author

Koen Derks (maintainer, author)	<k.derks@nyenrode.nl>

Please use the citation provided by R when citing this package. A BibTex entry is available from citation('jfa').

See also

Examples

# Load the jfa package
library(jfa)

# Load the BuildIt population
data('BuildIt')

############################################
### Example 1: Classical audit sampling ####
############################################

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01, 
                   likelihood = 'poisson', conf.level = 0.95)
summary(stage1)
#> 
#> 	Classical Audit Sample Planning Summary
#> 
#> Options:
#>   Confidence level:              0.95 
#>   Materiality:                   0.03 
#>   Hypotheses:                    H₀: Θ >= 0.03 vs. H₁: Θ < 0.03 
#>   Expected:                      0.01 
#>   Likelihood:                    poisson 
#> 
#> Results:
#>   Minimum sample size:           220 
#>   Tolerable errors:              2.2 
#>   Expected most likely error:    0.01 
#>   Expected upper bound:          0.02997 
#>   Expected precision:            0.01997 
#>   Expected p-value:              < 2.22e-16 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = 'values', values = 'bookValue',
                    method = 'interval', start = 1)
summary(stage2)
#> 
#> 	Audit Sample Selection Summary
#> 
#> Options:
#>   Requested sample size:         220 
#>   Sampling units:                monetary units 
#>   Method:                        fixed interval sampling 
#>   Starting point:                1 
#> 
#> Data:
#>   Population size:               3500 
#>   Population value:              1403221 
#>   Selection interval:            6378.3 
#> 
#> Results:
#>   Selected sampling units:       220 
#>   Proportion of value:           0.080554 
#>   Selected items:                220 
#>   Proportion of size:            0.062857 

# Stage 3: Execution
sample <- stage2[['sample']]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, method = 'stringer',
                     conf.level = 0.95, data = sample,
                     values = 'bookValue', values.audit = 'auditValue')
summary(stage4)
#> 
#> 	Classical Audit Sample Evaluation Summary
#> 
#> Options:
#>   Confidence level:               0.95 
#>   Materiality:                    0.03 
#>   Method:                         stringer 
#> 
#> Data:
#>   Sample size:                    220 
#>   Number of errors:               5 
#>   Sum of taints:                  2.9999929 
#> 
#> Results:
#>   Most likely error:              0.013636 
#>   95 percent confidence interval: [0, 0.033724] 
#>   Precision:                      0.020087 

######################################################################
### Example 2: Bayesian audit sampling using a non-informed prior ####
######################################################################

# Create the prior distribution
prior <- auditPrior(method = 'default', likelihood = 'poisson')
summary(prior)
#> 
#> 	Prior Distribution Summary
#> 
#> Options:
#>   Likelihood:                    poisson 
#>   Specifics:                     noninformative 
#> 
#> Results:
#>   Functional form:               gamma(α = 1, β = 1) 
#>   Equivalent sample size:        1 
#>   Equivalent errors:             0 
#>   Mode:                          0 
#>   Mean:                          1 
#>   Median:                        0.69315 
#>   Variance:                      1 
#>   Skewness:                      2 
#>   95 percent upper bound:        2.9957 
#>   Precision:                     2.9957 

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01,
                   likelihood = 'poisson', conf.level = 0.95, prior = prior)
summary(stage1)
#> 
#> 	Bayesian Audit Sample Planning Summary
#> 
#> Options:
#>   Confidence level:              0.95 
#>   Materiality:                   0.03 
#>   Hypotheses:                    H₀: Θ > 0.03 vs. H₁: Θ < 0.03 
#>   Expected:                      0.01 
#>   Likelihood:                    poisson 
#>   Prior distribution:            gamma(α = 1, β = 1) 
#> 
#> Results:
#>   Minimum sample size:           218 
#>   Tolerable errors:              2.18 
#>   Posterior distribution:        gamma(α = 3.18, β = 219) 
#>   Expected most likely error:    0.0099543 
#>   Expected upper bound:          0.029971 
#>   Expected precision:            0.020017 
#>   Expected BF₁₀:	                 626.86 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = 'values', values = 'bookValue',
                    method = 'interval', start = 1)
summary(stage2)
#> 
#> 	Audit Sample Selection Summary
#> 
#> Options:
#>   Requested sample size:         218 
#>   Sampling units:                monetary units 
#>   Method:                        fixed interval sampling 
#>   Starting point:                1 
#> 
#> Data:
#>   Population size:               3500 
#>   Population value:              1403221 
#>   Selection interval:            6436.8 
#> 
#> Results:
#>   Selected sampling units:       218 
#>   Proportion of value:           0.079411 
#>   Selected items:                218 
#>   Proportion of size:            0.062286 

# Stage 3: Execution
sample <- stage2[['sample']]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, conf.level = 0.95, data = sample,
                     values = 'bookValue', values.audit = 'auditValue',
                     prior = prior)
summary(stage4)
#> 
#> 	Bayesian Audit Sample Evaluation Summary
#> 
#> Options:
#>   Confidence level:               0.95 
#>   Materiality:                    0.03 
#>   Materiality:                    0.03 
#>   Hypotheses:                     H₀: Θ > 0.03 vs. H₁: Θ < 0.03 
#>   Method:                         poisson 
#>   Prior distribution:             gamma(α = 1, β = 1) 
#> 
#> Data:
#>   Sample size:                    218 
#>   Number of errors:               4 
#>   Sum of taints:                  2.4000172 
#> 
#> Results:
#>   Posterior distribution:         gamma(α = 3.4, β = 219) 
#>   Most likely error:              0.010959 
#>   95 percent credible interval:   [0, 0.03145] 
#>   Precision:                      0.020491 
#>   BF₁₀:	                          493.5 

###################################################################
### Example 3: Bayesian audit sampling using an informed prior ####
###################################################################

# Create the prior distribution
prior <- auditPrior(method = 'arm', likelihood = 'poisson',
                    expected = 0.01, materiality = 0.03, cr = 0.6, ir = 1)
summary(prior)
#> 
#> 	Prior Distribution Summary
#> 
#> Options:
#>   Likelihood:                    poisson 
#>   Specifics:                     ir = 1; cr = 0.6; dr = 0.0833333 
#> 
#> Results:
#>   Functional form:               gamma(α = 1.46, β = 46) 
#>   Equivalent sample size:        46 
#>   Equivalent errors:             0.46 
#>   Mode:                          0.01 
#>   Mean:                          0.031739 
#>   Median:                        0.024859 
#>   Variance:                      0.00069 
#>   Skewness:                      1.6552 
#>   95 percent upper bound:        0.08343 
#>   Precision:                     0.07343 

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01,
                   likelihood = 'poisson', conf.level = 0.95, prior = prior)
summary(stage1)
#> 
#> 	Bayesian Audit Sample Planning Summary
#> 
#> Options:
#>   Confidence level:              0.95 
#>   Materiality:                   0.03 
#>   Hypotheses:                    H₀: Θ > 0.03 vs. H₁: Θ < 0.03 
#>   Expected:                      0.01 
#>   Likelihood:                    poisson 
#>   Prior distribution:            gamma(α = 1.46, β = 46) 
#> 
#> Results:
#>   Minimum sample size:           174 
#>   Tolerable errors:              1.74 
#>   Posterior distribution:        gamma(α = 3.2, β = 220) 
#>   Expected most likely error:    0.01 
#>   Expected upper bound:          0.02997 
#>   Expected precision:            0.01997 
#>   Expected BF₁₀:	                 13.593 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = 'values', values = 'bookValue',
                    method = 'interval', start = 1)
summary(stage2)
#> 
#> 	Audit Sample Selection Summary
#> 
#> Options:
#>   Requested sample size:         174 
#>   Sampling units:                monetary units 
#>   Method:                        fixed interval sampling 
#>   Starting point:                1 
#> 
#> Data:
#>   Population size:               3500 
#>   Population value:              1403221 
#>   Selection interval:            8064.5 
#> 
#> Results:
#>   Selected sampling units:       174 
#>   Proportion of value:           0.058451 
#>   Selected items:                174 
#>   Proportion of size:            0.049714 

# Stage 3: Execution
sample <- stage2[['sample']]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, conf.level = 0.95, data = sample,
                     values = 'bookValue', values.audit = 'auditValue',
                     prior = prior)
summary(stage4)
#> 
#> 	Bayesian Audit Sample Evaluation Summary
#> 
#> Options:
#>   Confidence level:               0.95 
#>   Materiality:                    0.03 
#>   Materiality:                    0.03 
#>   Hypotheses:                     H₀: Θ > 0.03 vs. H₁: Θ < 0.03 
#>   Method:                         poisson 
#>   Prior distribution:             gamma(α = 1.46, β = 46) 
#> 
#> Data:
#>   Sample size:                    174 
#>   Number of errors:               4 
#>   Sum of taints:                  2.4000247 
#> 
#> Results:
#>   Posterior distribution:         gamma(α = 3.86, β = 220) 
#>   Most likely error:              0.013 
#>   95 percent credible interval:   [0, 0.034335] 
#>   Precision:                      0.021335 
#>   BF₁₀:	                          6.9023