jfa --- Bayesian and Classical Audit Sampling

jfa is an R package for statistical audit sampling. The package provides functions for planning, performing, evaluating, and reporting an audit sample. Specifically, these functions implement standard audit sampling techniques for calculating sample sizes, selecting items from a population, and evaluating the misstatement from a data sample or from summary statistics. Additionally, the jfa package allows the user to create a prior probability distribution to perform Bayesian audit sampling using these functions.

The package and its intended workflow are also implemented with a graphical user interface in the Audit module of JASP, a free and open-source statistical software program.

For documentation on jfa itself, including the manual and user guide for the package, worked examples, and other tutorial information visit the package website.

Reference tables

Below you can find several links to reference tables that contain statistical sample sizes, upper limits, and Bayes factors. These tables are created using the planning() and evaluation() functions provided in the package. See the corresponding help files for more information about these functions and how to replicate this output.

Sample sizes

Sample sizes based on the binomial distribution

Sample sizes based on the Poisson distribution

Sample sizes based on the hypergeometric distribution

Upper limits

Upper limits based on the binomial distribution

Upper limits based on the Poisson distribution

Upper limits based on the hypergeometric distribution

Bayes factors

Bayes factors based on the beta distribution

Bayes factors based on the gamma distribution

Bayes factors based on the beta-binomial distribution

Author

Koen Derks (maintainer, author)	<k.derks@nyenrode.nl>

Please use the citation provided by R when citing this package. A BibTex entry is available from citation("jfa").

See also

Examples

# Load the jfa package
library(jfa)

# Load the BuildIt population
data("BuildIt")

############################################
### Example 1: Classical audit sampling ####
############################################

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01, 
                   likelihood = "poisson", conf.level = 0.95)
summary(stage1)

#> # ------------------------------------------------------------ 
#> #              jfa Planning Summary (Classical)
#> # ------------------------------------------------------------ 
#> # Input:
#> # 
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Likelihood:                    poisson  
#> # Population size:               Not specified 
#> # Expected sample errors:        2.2 
#> # ------------------------------------------------------------
#> # Output:
#> #
#> # Sample size:                   220  
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Expected most likely error:    1% 
#> # Expected upper bound:          2.997% 
#> # Expected precision:            1.997% 
#> # Expected p value:              0.0498 
#> # ------------------------------------------------------------ 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = "values", values = "bookValue",
                    method = "interval", start = 1)
summary(stage2)

#> # ------------------------------------------------------------
#> #                  jfa Selection Summary
#> # ------------------------------------------------------------
#> # Input:
#> #      
#> # Population size:               1403221 
#> # Requested sample size:         220 
#> # Sampling units:                Monetary units 
#> # Method:                        Fixed interval sampling 
#> # Interval:                      6378.276 
#> # Starting point:                1 
#> # ------------------------------------------------------------ 
#> # Output:
#> #
#> # Obtained sampling units:       220 
#> # Obtained items:                220 
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Proportion n/N:                0.063 
#> # Percentage of value:           0.016% 
#> # ------------------------------------------------------------ 

# Stage 3: Execution
sample <- stage2[["sample"]]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, method = "stringer",
                     conf.level = 0.95, data = sample,
                     values = "bookValue", values.audit = "auditValue")
summary(stage4)

#> # ------------------------------------------------------------ 
#> #             jfa Evaluation Summary (Classical)
#> # ------------------------------------------------------------ 
#> # Input:
#> #
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Population size:               Not specified 
#> # Sample size:                   220 
#> # Sample errors:                 2 
#> # Sum of taints:                 1.2 
#> # Method:                        stringer   
#> # ------------------------------------------------------------
#> # Output:
#> #  
#> # Most likely error:             0.545%
#> # Upper bound:                   2.241%
#> # Precision:                     1.696% 
#> # p Value:                       Requires likelihood 
#> # Sufficient evidence:           TRUE 
#> # ------------------------------------------------------------ 

######################################################################
### Example 2: Bayesian audit sampling using a non-informed prior ####
######################################################################

# Create the prior distribution
prior <- auditPrior(method = "none", likelihood = "poisson")
summary(prior)

#> # ------------------------------------------------------------
#> #         jfa Prior Distribution Summary (Bayesian)
#> # ------------------------------------------------------------
#> # Input:
#> #
#> # Confidence level:              95% 
#> # Expected sample errors:        0% 
#> # Likelihood:                    poisson 
#> # Specifics:                     None 
#> # ------------------------------------------------------------
#> # Output: 
#> #
#> # Prior distribution:            gamma(α = 1, β = 0) 
#> # Implicit sample size:          0 
#> # Implicit errors:               0 
#> # ------------------------------------------------------------
#> # Statistics: 
#> #
#> # Upper bound:                   Inf 
#> # Precision:                     NaN 
#> # Mode:                          NaN 
#> # Mean:                          Inf 
#> # Median:                        Inf 
#> # ------------------------------------------------------------

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01,
                   likelihood = "poisson", conf.level = 0.95, prior = prior)
summary(stage1)

#> # ------------------------------------------------------------ 
#> #              jfa Planning Summary (Bayesian)
#> # ------------------------------------------------------------ 
#> # Input:
#> # 
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Likelihood:                    poisson  
#> # Population size:               Not specified 
#> # Expected sample errors:        2.2 
#> # ------------------------------------------------------------
#> # Output:
#> #
#> # Sample size:                   220  
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Expected most likely error:    1% 
#> # Expected upper bound:          2.997% 
#> # Expected precision:            1.997% 
#> # Expected Bayes factor-+:       19.096 
#> # ------------------------------------------------------------ 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = "values", values = "bookValue",
                    method = "interval", start = 1)
summary(stage2)

#> # ------------------------------------------------------------
#> #                  jfa Selection Summary
#> # ------------------------------------------------------------
#> # Input:
#> #      
#> # Population size:               1403221 
#> # Requested sample size:         220 
#> # Sampling units:                Monetary units 
#> # Method:                        Fixed interval sampling 
#> # Interval:                      6378.276 
#> # Starting point:                1 
#> # ------------------------------------------------------------ 
#> # Output:
#> #
#> # Obtained sampling units:       220 
#> # Obtained items:                220 
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Proportion n/N:                0.063 
#> # Percentage of value:           0.016% 
#> # ------------------------------------------------------------ 

# Stage 3: Execution
sample <- stage2[["sample"]]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, conf.level = 0.95, data = sample,
                     values = "bookValue", values.audit = "auditValue",
                     prior = prior)
summary(stage4)

#> # ------------------------------------------------------------ 
#> #             jfa Evaluation Summary (Bayesian)
#> # ------------------------------------------------------------ 
#> # Input:
#> #
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Population size:               Not specified 
#> # Sample size:                   220 
#> # Sample errors:                 2 
#> # Sum of taints:                 1.2 
#> # Method:                        poisson  
#> # Prior distribution:            gamma(α = 1, β = 0) 
#> # ------------------------------------------------------------
#> # Output:
#> # 
#> # Posterior distribution:        gamma(α = 2.2, β = 220) 
#> # Most likely error:             0.545%
#> # Upper bound:                   2.302%
#> # Precision:                     1.757% 
#> # Bayes factor-+:                69.896 
#> # Sufficient evidence:           TRUE 
#> # ------------------------------------------------------------ 

###################################################################
### Example 3: Bayesian audit sampling using an informed prior ####
###################################################################

# Create the prior distribution
prior <- auditPrior(method = "arm", likelihood = "poisson",
                    expected = 0.01, materiality = 0.03, cr = 0.6, ir = 1)
summary(prior)

#> # ------------------------------------------------------------
#> #         jfa Prior Distribution Summary (Bayesian)
#> # ------------------------------------------------------------
#> # Input:
#> #
#> # Confidence level:              95% 
#> # Expected sample errors:        1% 
#> # Likelihood:                    poisson 
#> # Specifics:                     IR = 1; ICR = 0.6; DR = 0.083 
#> # ------------------------------------------------------------
#> # Output: 
#> #
#> # Prior distribution:            gamma(α = 1.46, β = 46) 
#> # Implicit sample size:          46 
#> # Implicit errors:               0.46 
#> # ------------------------------------------------------------
#> # Statistics: 
#> #
#> # Upper bound:                   0.083 
#> # Precision:                     0.073 
#> # Mode:                          0.01 
#> # Mean:                          0.032 
#> # Median:                        0.025 
#> # ------------------------------------------------------------

# Stage 1: Planning
stage1 <- planning(materiality = 0.03, expected = 0.01,
                   likelihood = "poisson", conf.level = 0.95, prior = prior)
summary(stage1)

#> # ------------------------------------------------------------ 
#> #              jfa Planning Summary (Bayesian)
#> # ------------------------------------------------------------ 
#> # Input:
#> # 
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Likelihood:                    poisson  
#> # Population size:               Not specified 
#> # Expected sample errors:        1.74 
#> # ------------------------------------------------------------
#> # Output:
#> #
#> # Sample size:                   174  
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Expected most likely error:    1% 
#> # Expected upper bound:          2.997% 
#> # Expected precision:            1.997% 
#> # Expected Bayes factor-+:       13.593 
#> # ------------------------------------------------------------ 

# Stage 2: Selection
stage2 <- selection(data = BuildIt, size = stage1,
                    units = "values", values = "bookValue",
                    method = "interval", start = 1)
summary(stage2)

#> # ------------------------------------------------------------
#> #                  jfa Selection Summary
#> # ------------------------------------------------------------
#> # Input:
#> #      
#> # Population size:               1403221 
#> # Requested sample size:         174 
#> # Sampling units:                Monetary units 
#> # Method:                        Fixed interval sampling 
#> # Interval:                      8064.487 
#> # Starting point:                1 
#> # ------------------------------------------------------------ 
#> # Output:
#> #
#> # Obtained sampling units:       174 
#> # Obtained items:                174 
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Proportion n/N:                0.05 
#> # Percentage of value:           0.012% 
#> # ------------------------------------------------------------ 

# Stage 3: Execution
sample <- stage2[["sample"]]

# Stage 4: Evaluation
stage4 <- evaluation(materiality = 0.03, conf.level = 0.95, data = sample,
                     values = "bookValue", values.audit = "auditValue",
                     prior = prior)
summary(stage4)

#> # ------------------------------------------------------------ 
#> #             jfa Evaluation Summary (Bayesian)
#> # ------------------------------------------------------------ 
#> # Input:
#> #
#> # Confidence level:              95% 
#> # Materiality:                   3% 
#> # Minimum precision:             Not specified 
#> # Population size:               Not specified 
#> # Sample size:                   174 
#> # Sample errors:                 0 
#> # Sum of taints:                 0 
#> # Method:                        poisson  
#> # Prior distribution:            gamma(α = 1.46, β = 46) 
#> # ------------------------------------------------------------
#> # Output:
#> # 
#> # Posterior distribution:        gamma(α = 1.46, β = 220) 
#> # Most likely error:             0.209%
#> # Upper bound:                   1.744%
#> # Precision:                     1.535% 
#> # Bayes factor-+:                181.942 
#> # Sufficient evidence:           TRUE 
#> # ------------------------------------------------------------