Frequentist and Bayesian Planning for Audit Sampling

This function calculates the required sample size for an audit, based on the Poisson, binomial or hypergeometric likelihood. A prior distribution can be specified to perform Bayesian planning. The returned object is of class jfaPlanning and has a print() and plot() method.

For more details on how to use this function see the package vignette: vignette('jfa', package = 'jfa')

planning(confidence = 0.95, expectedError = 0, likelihood = 'poisson', N = NULL, 
          materiality = NULL, minPrecision = NULL, 
          prior = FALSE, nPrior = 0, kPrior = 0,
          increase = 1, maxSize = 5000)

Arguments

confidence
a value between 0 and 1 specifying the confidence level desired for the sample planning. Defaults to 0.95 for 95% confidence.

expectedError
a value between 0 and 1 specifying the expected errors in the sample relative to the total sample size, or a number (>= 1) that represents the number of expected errors in the sample. It is advised to set this value conservatively to minimize the probability of the observed errors exceeding the expected errors, which would imply that insufficient work has been done in the end.

likelihood
a character specifying the likelihood assumed in the calculation. This can be either binomial for the binomial likelihood, poisson for the Poisson likelihood, or hypergeometric for the hypergeometric likelihood. See the details section for more information about the available likelihoods.

N
an integer (or value) larger than 0 specifying the total population size (or value). Only required when likelihood = 'hypergeometric'.

materiality
a value between 0 and 1 specifying the performance materiality (i.e., maximum upper limit) of the audit as a fraction of the total size (or value). Can be NULL, but minPrecision should be specified in that case.

minPrecision
a value between 0 and 1 specifying the minimum precision (i.e., upper bound minus most likely error) to be obtained. Can be NULL, but materiality should be specified in that case.

prior
a logical specifying whether to use a prior distribution when planning, or an object of class `jfaPrior` containing the prior distribution. Defaults to FALSE for frequentist planning. If TRUE, a negligible prior distribution is chosen by default, but can be adjusted using the `kPrior` and `nPrior` arguments. Chooses a conjugate gamma distribution for the Poisson likelihood, a conjugate beta distribution for the binomial likelihood, and a conjugate beta-binomial distribution for the hypergeometric likelihood.

nPrior
if prior = TRUE, a value larger than, or equal to, 0 specifying the number of sampling units in the implicit sample on which the prior distribution is based.

kPrior
if prior = TRUE, a value larger than, or equal to, 0 specifying the assumed sum of errors in the implicit sample on which the prior distribution is based.

increase
an integer larger than 0 specifying the desired increase step for the sample size calculation.

maxSize
an integer larger than 0 specifying the maximum sample size that is considered in the calculation. Defaults to 5000 for efficiency. Increase this value if the sample size cannot be found due to it being too large (e.g., for a low materiality).

Value

An object of class jfaPlanning containing:

confidence

a value between 0 and 1 indicating the confidence level.

expectedError

a value larger than, or equal to, 0 indicating the expected errors input.

likelihood

a character indicating the specified likelihood.

N

an integer larger than 0 indicating the population size (only returned if N is specified).

materiality

a value between 0 and 1 indicating the specified materiality. Can be NULL.

minPrecision

a value between 0 and 1 indicating the minimum precision to be obtained. Can be NULL.

sampleSize

an integer larger than 0 indicating the required sample size.

errorType

a character indicating whether the expected errors where specified as a percentage or as an integer.

expectedSampleError

a value larger than, or equal to, 0 indicating the number of errors that are allowed in the sample.

expectedBound

a value between 0 and 1 indicating the expected upper bound if the sample goes according to plan.

expectedPrecision

a value between 0 and 1 indicating the expected precision if the sample goes according to plan.

populationK

if likelihood = 'hypergeometric', an integer larger than 0 indicating the assumed population errors.

prior

if a prior distribution is specified, an object of class jfaPrior that contains the prior distribution.

expectedPosterior

if a prior distribution is specified, an object of class jfaPosterior that contains the expected posterior distribution.

Details

This section elaborates on the available likelihoods and corresponding prior distributions for the likelihood argument.

poisson: The Poisson likelihood is used as a likelihood for monetary unit sampling (MUS). Its likelihood function is defined as: $$p(x) = \frac{\lambda^x e^{-\lambda}}{x!}$$ The conjugate gamma(\(\alpha, \beta\)) prior has probability density function: $$f(x; \alpha, \beta) = \frac{\beta^\alpha x^{\alpha - 1} e^{-\beta x}}{\Gamma(\alpha)}$$

binomial: The binomial likelihood is used as a likelihood for record sampling with replacement. Its likelihood function is defined as: $$p(x) = {n \choose k} p^k (1 - p)^{n - k}$$ The conjugate beta(\(\alpha, \beta\)) prior has probability density function: $$f(x; \alpha, \beta) = \frac{1}{B(\alpha, \beta)} x^{\alpha - 1} (1 - x)^{\beta - 1}$$

hypergeometric: The hypergeometric likelihood is used as a likelihood for record sampling without replacement. Its likelihood function is defined as: $$p(x = k) = \frac{{K \choose k} {N - K \choose n - k}}{{N \choose n}}$$ The conjugate beta-binomial(\(\alpha, \beta\)) prior (Dyer and Pierce, 1993) has probability density function: $$f(k | n, \alpha, \beta) = {n \choose k} \frac{B(k + \alpha, n - k + \beta)}{B(\alpha, \beta)}$$

References

Dyer, D. and Pierce, R.L. (1993). On the choice of the prior distribution in hypergeometric sampling. Communications in Statistics - Theory and Methods, 22(8), 2125 - 2146.

See also

Author

Koen Derks, k.derks@nyenrode.nl

Examples

library(jfa)

# Using the binomial distribution, calculates the required sample size for a 
# materiality of 5% when 2.5% mistakes are expected to be found in the sample.

# Frequentist planning with binomial likelihood:

p1 <- planning(confidence = 0.95, expectedError = 0.025, likelihood = 'binomial',
               materiality = 0.05)
print(p1)

#> # ------------------------------------------------------------
#> #              jfa Planning Summary (Frequentist)
#> # ------------------------------------------------------------     
#> # Input:
#> # 
#> # Confidence:              95% 
#> # Materiality:             5% 
#> # Minimum precision:       Not specified 
#> # Likelihood:              binomial 
#> # Expected sample errors:  6 
#> # ------------------------------------------------------------
#> # Output:
#> #
#> # Sample size:             234 
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Expected upper bound:    5% 
#> # Expected precision:      2.43% 
#> # ------------------------------------------------------------ 

# ------------------------------------------------------------
#              jfa Planning Summary (Frequentist)
# ------------------------------------------------------------     
# Input:
# 
# Confidence:              95% 
# Materiality:             5% 
# Minimum precision:       Not specified 
# Likelihood:              binomial 
# Expected sample errors:  6 
# ------------------------------------------------------------
# Output:
#
# Sample size:             234 
# ------------------------------------------------------------
# Statistics:
#
# Expected upper bound:    5% 
# Expected precision:      2.43% 
# ------------------------------------------------------------ 

# Bayesian planning with prior:

prior <- auditPrior(confidence = 0.95, likelihood = 'binomial', method = 'arm', 
                    expectedError = 0.025, materiality = 0.05, cr = 0.6)

p2 <- planning(confidence = 0.95, expectedError = 0.025, materiality = 0.05,
               prior = prior)
print(p2)

#> # ------------------------------------------------------------
#> #              jfa Planning Summary (Bayesian)
#> # ------------------------------------------------------------
#> # Input:
#> # 
#> # Confidence:              95%      
#> # Materiality:             5% 
#> # Minimum precision:       Not specified 
#> # Likelihood:              binomial 
#> # Prior distribution:      beta(α = 2.275, β = 50.725) 
#> # Expected sample errors:  4.23 
#> # ------------------------------------------------------------
#> # Output:
#> #
#> # Sample size:             169 
#> # Posterior distribution:  beta(α = 6.5, β = 215.5) 
#> # ------------------------------------------------------------
#> # Statistics:
#> #
#> # Expected upper bound:    4.99% 
#> # Expected precision:      2.49% 
#> # Expected Bayes factor-+: 9.32 
#> # ------------------------------------------------------------ 

# ------------------------------------------------------------
#              jfa Planning Summary (Bayesian)
# ------------------------------------------------------------
# Input:
# 
# Confidence:              95%      
# Materiality:             5% 
# Minimum precision:       Not specified 
# Likelihood:              binomial 
# Prior distribution:      beta(2.275, 50.725) 
# Expected sample errors:  4.23 
# ------------------------------------------------------------
# Output:
#
# Sample size:             169 
# Posterior distribution:  beta(6.5, 215.5) 
# ------------------------------------------------------------
# Statistics:
#
# Expected upper bound:    4.99% 
# Expected precision:      2.49% 
# Expected Bayes factor-+: 9.32 
# ------------------------------------------------------------